Move file loader unit test to its own file
